Support wrapping multi-line nested elements
This is the case where an argument, list element, etc. within
the span we're wrapping already has some wrapping of its own.
We don't want to change the shape of that indentation, but we
do need to make it more indented such that it lines up with
the rest of the new layout.