Cope with positions being specified in any order
Our edits rely on being specified in order of their position in
the document to be sure that they apply cleanly. Since we can't
control the input order of the positions we need to enforce this
ordering on the resulting edits.

While we could do this at the top (i.e: sort the input positions)
and rely on ordering throughout, sorting at the bottom allows
greater freedom for the intervening algorithm.